+++ src/app/(afterLogin)/layout.tsx
@@ -21,14 +21,15 @@ export default function AfterLoginlayout({
                   </ul>
                 </nav>
               </div>
+              {/* <div className="mt-10 mb-1 bg-slate-950 w-1 ">
+                <Divider orientation="vertical" />
+              </div> */}
             </section>
           </header>
-          <div className="mt-10 mb-1 bg-slate-950 w-1">
-            <Divider orientation="vertical" />
-          </div>
+
           <div className="mt-10 mb-2 flex items-start flex-col flex-grow">
             <div className="w-full h-full flex justify-between">
-              <main className="w-full p-10">{children}</main>
+              <main className="w-full ">{children}</main>
             </div>
           </div>
         </RQProvider>
